Unify header files for ttrace. It should be in os/include/tinyara folder because ttrace is a special feature, not common for TizenRT
