+++ lib/libc/ttrace/lib_ttrace.c
@@ -23,16 +23,17 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <string.h>
-#include <ttrace.h>
 #include <sys/ioctl.h>
 #include <sys/types.h>
 #include <tinyara/clock.h>
-#include <tinyara/ttrace_internal.h>
+#include <tinyara/ttrace.h>
 #include <tinyara/sched.h>
 
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
+#define TTRACE_EVENT_TYPE_BEGIN    'b'
+#define TTRACE_EVENT_TYPE_END      'e'
 
 /****************************************************************************
  * Private Type Declarations
@@ -66,7 +67,7 @@ int fd = -1;
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
-int is_fd_available(void)
+static int is_fd_available(void)
 {
 	if (fd < 0) {
 		fd = open("/dev/ttrace", O_WRONLY);
@@ -75,19 +76,63 @@ int is_fd_available(void)
 	return fd;
 }
 
-int is_tag_available(int tag)
+static bool is_tag_available(int tag)
 {
-	return ioctl(fd, TTRACE_FUNC_TAG, tag);
+	int ret;
+
+	ret = ioctl(fd, TTRACE_FUNC_TAG, tag);
+	if (!(ret & tag)) {
+		return false;
+	}
+
+	return true;
 }
 
-int send_packet_sched(struct trace_packet *packet)
+static int show_packet(struct trace_packet *packet)
+{
+	int uid = (packet->codelen & TTRACE_CODE_UNIQUE) >> 7;
+	int msg_len = (packet->codelen & ~TTRACE_CODE_UNIQUE);
+	int pad = 0;
+	ttdbg("time: %06d.%06d\r\n", packet->ts.tv_sec, packet->ts.tv_usec);
+	ttdbg("event_type: %c, %d\r\n", packet->event_type, packet->event_type);
+	ttdbg("pid: %d\r\n", packet->pid);
+	ttdbg("codelen: %d\r\n", packet->codelen);
+	ttdbg("unique code? %d\r\n", uid);
+	if (uid == TRUE) {
+		ttdbg("uid: %d\r\n", (packet->codelen & ~TTRACE_CODE_UNIQUE));
+		pad = TTRACE_MSG_BYTES;
+	} else {
+		ttdbg("message: %s\r\n", packet->msg.message);
+		pad = 0;
+	}
+
+	return sizeof(struct trace_packet) - pad;
+}
+
+static int show_sched_packet(struct trace_packet *packet)
+{
+	ttdbg("[%06d:%06d] %03d: %c|prev_comm=%s prev_pid=%u prev_prio=%u prev_state=%u ==> next_comm=%s next_pid=%u next_prio=%u\r\n",
+		  packet->ts.tv_sec, packet->ts.tv_usec,
+		  packet->pid,
+		  (char)packet->event_type,
+		  packet->msg.sched_msg.prev_comm,
+		  packet->msg.sched_msg.prev_pid,
+		  packet->msg.sched_msg.prev_prio,
+		  packet->msg.sched_msg.prev_state,
+		  packet->msg.sched_msg.next_comm,
+		  packet->msg.sched_msg.next_pid,
+		  packet->msg.sched_msg.next_prio);
+	return sizeof(struct trace_packet);
+}
+
+static int send_packet_sched(struct trace_packet *packet)
 {
 	int ret = 0;
 	ret = write(fd, packet, sizeof(struct trace_packet));
 	return ret;
 }
 
-int create_packet_sched(struct trace_packet *packet, struct tcb_s *prev, struct tcb_s *next)
+static int create_packet_sched(struct trace_packet *packet, struct tcb_s *prev, struct tcb_s *next)
 {
 	int ret = TTRACE_VALID;
 	int msg_len = sizeof(struct sched_message);
@@ -127,7 +172,7 @@ int create_packet_sched(struct trace_packet *packet, struct tcb_s *prev, struct 
 	return ret;
 }
 
-int send_packet(struct trace_packet *packet)
+static int send_packet(struct trace_packet *packet)
 {
 	int ret = 0;
 
@@ -140,7 +185,7 @@ int send_packet(struct trace_packet *packet)
 	return ret;
 }
 
-int create_packet(struct trace_packet *packet, char type, char *str, va_list valist)
+static int create_packet(struct trace_packet *packet, char type, char *str, va_list valist)
 {
 	int ret = TTRACE_VALID;
 	int msg_len = strlen(str);
@@ -160,7 +205,7 @@ int create_packet(struct trace_packet *packet, char type, char *str, va_list val
 	return ret;
 }
 
-int create_packet_u(struct trace_packet *packet, char type, int8_t uid)
+static int create_packet_u(struct trace_packet *packet, char type, int8_t uid)
 {
 	int ret = 0;
 	gettimeofday(&(packet->ts), NULL);
@@ -174,7 +219,6 @@ int create_packet_u(struct trace_packet *packet, char type, int8_t uid)
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
-
 int trace_sched(struct tcb_s *prev_tcb, struct tcb_s *next_tcb)
 {
 	int ret = TTRACE_VALID;
@@ -200,6 +244,7 @@ int trace_sched(struct tcb_s *prev_tcb, struct tcb_s *next_tcb)
 	return ret;
 
 }
+
 /****************************************************************************
  * Name: trace_begin
  *
