+++ os/kernel/pthread/pthread_detach.c
@@ -62,11 +62,11 @@
 #include <errno.h>
 #include <assert.h>
 #include <debug.h>
+#include <tinyara/ttrace.h>
 
 #include "sched/sched.h"
 #include "group/group.h"
 #include "pthread/pthread.h"
-#include <ttrace.h>
 
 /************************************************************************
  * Pre-processor Definitions
