+++ os/kernel/task/task_prctl.c
@@ -63,10 +63,10 @@
 #include <debug.h>
 
 #include <tinyara/sched.h>
+#include <tinyara/ttrace.h>
 
 #include "sched/sched.h"
 #include "task/task.h"
-#include <ttrace.h>
 
 /************************************************************************
  * Private Functions
